requirements.txt
amqp==2.2.1
asgiref==3.2.3
billiard==3.5.0.3
celery==4.1.0
Django==3.0.2
django-crispy-forms==1.8.1
django-taggit==1.1.0
django-tinymce4-lite==1.7.5
importlib-metadata==0.23
jsmin==2.2.2
kombu==4.1.0
Markdown==3.1.1
more-itertools==7.2.0
Pillow==7.0.0
psycopg2==2.8.3
pytz==2017.2
sqlparse==0.3.0
vine==1.1.4
virtualenv==16.7.6
zipp==0.6.0
